--- FaceRecognitionPanel/DBContext/Class.cs
@@ -1,10 +1,0 @@
-﻿namespace DBContext;
-
-public class Class
-{
-    public int ClassId { get; set; }
-
-    public string? ClassName { get; set; }
-
-    public virtual ICollection<User> Users { get; set; } = new List<User>();
-}
