+++ FaceRecognitionPanel/DBContext/Classi.cs
@@ -1,0 +1,8 @@
+namespace DBContext;
+
+public class Classi
+{
+    public int Id { get; set; }
+    public string Sezione { get; set; }
+    public virtual List<Utenti> utentis { get; set; }
+}
